Refactor bot and service scripts for improved readability and error handling
This commit enhances the structure of the bot and service scripts by organizing imports, improving logging, and refining error handling. The `start` and `getanime` functions have been updated for better message formatting and clarity. Additionally, the downloader and uploader services have been modified to handle exceptions more gracefully and ensure proper file management. Overall, these changes aim to improve code maintainability and user experience.

+++ downloaderService/main.py
@@ -4,11 +4,16 @@ import subprocess
 import os
 import sys
 
-def downloadVideo(search_query,search_query_range):
+def downloadVideo(search_query, search_query_range):
     # animdl download "demon slayer" -r 1
-    print('animdl', 'download',search_query,f'-r {str(search_query_range)}','--auto')
-    subprocess.run(['animdl', 'download',search_query,f'-r {str(search_query_range)}','--auto'])
-    
+    print('animdl', 'download', search_query,
+          f'-r {str(search_query_range)}', '--auto')
+    subprocess.run([
+        'animdl', 'download', search_query,
+        f'-r {str(search_query_range)}', '--auto'
+    ])
+
+
 def getalltsfiles():
     walk_dir = os.getcwd()
     for root, _, files in os.walk(walk_dir):
@@ -22,23 +27,30 @@ def getalltsfiles():
 
 
 def convert2mp4(infile, outfile):
-    # ffmpeg -i E01.ts -c:v copy -c:a copy -preset:v ultrafast -segment_list_flags +live video.mp4
-    subprocess.run(['ffmpeg','-i',infile,'-c:v','copy','-c:a','copy','-preset:v','ultrafast','-segment_list_flags','+live',outfile])
+    # ffmpeg -i E01.ts -c:v copy -c:a copy -preset:v ultrafast
+    # -segment_list_flags +live video.mp4
+    subprocess.run([
+        'ffmpeg', '-i', infile, '-c:v', 'copy', '-c:a', 'copy',
+        '-preset:v', 'ultrafast', '-segment_list_flags', '+live', outfile
+    ])
 
 
 def main(argv):
     search_query = argv[1]
     print(search_query)
     search_query_range = argv[2]
-    #anime_quality = argv[3] if len(sys.argv) >= 4 else '720[subtitle]'
+    # anime_quality = argv[3] if len(sys.argv) >= 4 else '720[subtitle]'
     search_query_in_quotes = f'"{search_query}"'
-    downloadVideo(search_query_in_quotes,search_query_range)
+    downloadVideo(search_query_in_quotes, search_query_range)
     try:
         infile, outfile = getalltsfiles()
-    except:
-        infile, outfile = None,None
-    convert2mp4(infile,outfile)
-    os.remove(infile)
+    except Exception:
+        infile, outfile = None, None
+
+    if infile and outfile:
+        convert2mp4(infile, outfile)
+        if os.path.exists(infile):
+            os.remove(infile)
